[README] Further clarify eligible versions
In short, only games played on the latest version are accepted.  This is
because it is highly likely that bugs in newer versions could affect
older versions of the map.  It quickly becomes very costly to view
replays on different versions while trying to be aware of every
applicable bug and issue.